Fix cilium network policy creation
